[GVN] Decouple GVNValueTable and GVNPass
`GVNValueTable` has several member functions taking a `GVNPass`
reference as an argument. This prevents moving `GVNPass`
out of `GVN.h` and into `GVN.cpp.`

The `GVNPass` reference is only used to access the `LeaderMap`. This patch moves
the `LeaderMap` type into its own header file, `GVNLeaderMap.h`, under class
name `GVNLeaderMap` and changes the `GVNValueTable` member functions to take a
`GVNLeaderMap` reference instead of a `GVNPass` reference.
